+++ src/filtering.py
@@ -41,3 +41,5 @@ def max_filt(s):
     return np.array(ss[::-1])
 
 
+
+