+++ src/knnsearch.py
@@ -239,3 +239,5 @@ def subset_nbrs(distance, indices, new_knn, block_size=100000):
     # Returning new_distance and new_indices as arrays.
     return np.asarray(new_distance), np.asarray(new_indices)
 
+
+