+++ src/reporting.py
@@ -1,10 +1,30 @@
-#!/usr/bin/env python2
-# -*- coding: utf-8 -*-
-"""
-Created on Mon Apr 15 14:39:28 2019
+# Copyright (c) 2019, Matheus Boni Vicari, treestruct
+# All rights reserved.
+#
+#
+#    This program is free software: you can redistribute it and/or modify
+#    it under the terms of the GNU General Public License as published by
+#    the Free Software Foundation, either version 3 of the License, or
+#    (at your option) any later version.
+#
+#    This program is distributed in the hope that it will be useful,
+#    but WITHOUT ANY WARRANTY; without even the implied warranty of
+#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#    GNU General Public License for more details.
+#
+#    You should have received a copy of the GNU General Public License
+#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-@author: matheus
-"""
+
+__author__ = "Matheus Boni Vicari"
+__copyright__ = "Copyright 2019, treestruct"
+__credits__ = ["Matheus Boni Vicari"]
+__license__ = "GPL3"
+__version__ = "0.11"
+__maintainer__ = "Matheus Boni Vicari"
+__email__ = "matheus.boni.vicari@gmail.com"
+__status__ = "Development"
+
 
 import os
 import numpy as np
@@ -28,7 +48,7 @@ def pdf_report(out_file, cloud_file, struct_file, ply_file):
     mlab.triangular_mesh(vertex_coords[:, 0], vertex_coords[:, 1],
                          vertex_coords[:, 2], facets_tri, scalars=vertex_ids)
     cam = f.scene.camera
-    cam.zoom(1.7)
+    cam.zoom(1.2)
     mlab.draw()
     mlab.savefig('temp_model.png')
     mlab.close()
@@ -71,3 +91,5 @@ def pdf_report(out_file, cloud_file, struct_file, ply_file):
     os.remove('temp_model.png')
         
     return
+
+